Kernel/Graphics: Apply DisplayConnector design on Intel driver
