+++ Kernel/Graphics/GraphicsManagement.cpp
@@ -14,8 +14,7 @@
 #include <Kernel/Graphics/GraphicsManagement.h>
 #include <Kernel/Graphics/Intel/NativeGraphicsAdapter.h>
 #include <Kernel/Graphics/VGA/ISAAdapter.h>
-#include <Kernel/Graphics/VGA/PCIAdapter.h>
-#include <Kernel/Graphics/VGACompatibleAdapter.h>
+#include <Kernel/Graphics/VGA/PCIGenericAdapter.h>
 #include <Kernel/Graphics/VirtIOGPU/GraphicsAdapter.h>
 #include <Kernel/Memory/AnonymousVMObject.h>
 #include <Kernel/Multiboot.h>
@@ -153,7 +152,7 @@ UNMAP_AFTER_INIT bool GraphicsManagement::determine_and_initialize_graphics_devi
             dmesgln("Graphics: The framebuffer set up by the bootloader is not RGB, ignoring fbdev argument");
         } else {
             dmesgln("Graphics: Using a preset resolution from the bootloader");
-            adapter = PCIVGACompatibleAdapter::initialize_with_preset_resolution(device_identifier,
+            adapter = PCIVGAGenericAdapter::must_create_with_preset_resolution(device_identifier,
                 multiboot_framebuffer_addr,
                 multiboot_framebuffer_width,
                 multiboot_framebuffer_height,
@@ -199,8 +198,8 @@ UNMAP_AFTER_INIT bool GraphicsManagement::determine_and_initialize_graphics_devi
             if (!m_vga_adapter && PCI::is_io_space_enabled(device_identifier.address())) {
                 create_bootloader_framebuffer_device();
             } else {
-                dmesgln("Graphics: Using a PCI VGA compatible generic adapter");
-                adapter = PCIVGACompatibleAdapter::initialize(device_identifier);
+                dmesgln("Graphics: Using a VGA compatible generic adapter");
+                adapter = PCIVGAGenericAdapter::must_create(device_identifier);
             }
             break;
         }
@@ -214,7 +213,7 @@ UNMAP_AFTER_INIT bool GraphicsManagement::determine_and_initialize_graphics_devi
     // Note: If no other VGA adapter is attached as m_vga_adapter, we should attach it then.
     if (!m_vga_adapter && PCI::is_io_space_enabled(device_identifier.address()) && adapter->vga_compatible()) {
         dbgln("Graphics adapter @ {} is operating in VGA mode", device_identifier.address());
-        m_vga_adapter = static_ptr_cast<VGACompatibleAdapter>(adapter);
+        m_vga_adapter = static_ptr_cast<VGAGenericAdapter>(adapter);
     }
     return true;
 }
